requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
plotly>=5.24.0
sodapy>=2.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
openpyxl>=3.1.0
